[TEMP] Adding iterate function to UtilsBase. Adding shell function to CommandsBase. Removing if __name__ = '__main__' statements to remove bloat. Minor change to __schedular function in SchedularBase
